Added code to ensure calcos will work under Python 3 and also under numpy 1.10.
git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/calcos/trunk@48905 fe389314-cf27-0410-b35b-8c050e845b92

+++ lib/calcos/xd_search.py
@@ -327,7 +327,7 @@ def findFwhm(e_j, y_locn):
     e_j_sorted = np.sort(e_j)
 
     third = nelem // 3
-    background = e_j_sorted[0:third].mean()
+    background = e_j_sorted[0:third].mean(dtype=np.float64)
 
     find_this_level = (e_max - background) / 2. + background
